NN-app/pyproject.toml
[project]
name = "nn-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch",
    "torchvision",
    "fastapi[standard]",
    "uvicorn", 
    "Pillow",
    "python-multipart",
]
